pyproject.toml
[project]
name = "invite-agent"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "fastapi>=0.118.0",
    "google-adk>=1.15.1",
    "gradio>=5.49.0",
    "psycopg2-binary>=2.9.11",
    "pydantic>=2.11.10",
    "python-dotenv>=1.1.1",
    "requests>=2.32.5",
    "uvicorn>=0.37.0",
    "passlib[bcrypt]>=1.7.4",
    "bcrypt==4.2.1",
    "pyjwt>=2.8.0",
    "asyncpg>=0.29.0",
    "pydantic[email]>=2.11.10",
]
